+++ codigofonte/ClientApp/app/shared2/localidade/localidade.service.ts
@@ -114,6 +114,21 @@ export class LocalidadeService2 {
         return this.getUfByCodigo(identificador);
     }
 
+    public getLocalidadeById(codigoLocalidade: number): Localidade{
+
+        if(codigoLocalidade == 0){
+
+            return this.getRoot();
+        }
+
+        if(codigoLocalidade.toString().length == 2){
+
+            return this.getUfByCodigo(codigoLocalidade);
+        }
+
+        return this.getMunicipioByCodigo(codigoLocalidade);
+    }
+
     public getUfByCodigo(ufCodigo: number) {
         if (!ufCodigo) return;
         return this._ufs.buscarPorCodigo(ufCodigo);
@@ -188,7 +203,7 @@ export class LocalidadeService2 {
     /**
      * Retorna a preposição (do, da ou de) mais adequada ao nome da UF.
      */
-    public getPreprosisaoTituloUF(nomeUF: string): string {
+    public getPreprosicaoTituloUF(nomeUF: string): string {
 
         let ufsComPreposicaoDo = ['acre', 'amapá', 'amazonas', 'ceará', 'distrito federal', 'mato grosso', 'mato grosso do sul', 'maranhão', 'paraná', 'pará', 'piauí', 'rio grande do norte', 'rio grande do sul', 'rio de janeiro'];
         let ufsComPreposicaoDa = ['bahia', 'paraíba'];